# fix: improve validation performance 5x