For a `log_metrics()` call with 100k fields, this brings down validation from ~250ms to ~50ms. Scales ~linearly: for 10k fields it's ~7ms. For 100 fields it's < 1ms.